Move `_choose_chains` to `backends.base`
