+++ pymc/sampling/mcmc.py
@@ -22,7 +22,7 @@ import warnings
 
 from collections import defaultdict
 from copy import copy
-from typing import Any, Dict, Iterator, List, Optional, Sequence, Tuple, Union, cast
+from typing import Any, Dict, Iterator, List, Optional, Sequence, Tuple, Union
 
 import aesara.gradient as tg
 import cloudpickle
@@ -34,7 +34,7 @@ from typing_extensions import TypeAlias
 
 import pymc as pm
 
-from pymc.backends.base import BaseTrace, MultiTrace
+from pymc.backends.base import BaseTrace, MultiTrace, _choose_chains
 from pymc.backends.ndarray import NDArray
 from pymc.blocking import DictToArrayBijection
 from pymc.exceptions import SamplingError
@@ -1478,33 +1478,6 @@ def log_warning_stats(stats: Sequence[Dict[str, Any]]):
         else:
             _log.warning(warn)
     return
-
-
-def _choose_chains(traces: Sequence[BaseTrace], tune: int) -> Tuple[List[BaseTrace], int]:
-    """
-    Filter and slice traces such that (n_traces * len(shortest_trace)) is maximized.
-
-    We get here after a ``KeyboardInterrupt``, and so the different
-    traces have different lengths. We therefore pick the number of
-    traces such that (number of traces) * (length of shortest trace)
-    is maximised.
-    """
-    if not traces:
-        raise ValueError("No traces to slice.")
-
-    lengths = [max(0, len(trace) - tune) for trace in traces]
-    if not sum(lengths):
-        raise ValueError("Not enough samples to build a trace.")
-
-    idxs = np.argsort(lengths)
-    l_sort = np.array(lengths)[idxs]
-
-    use_until = cast(int, np.argmax(l_sort * np.arange(1, l_sort.shape[0] + 1)[::-1]))
-    final_length = l_sort[use_until]
-
-    take_idx = cast(Sequence[int], idxs[use_until:])
-    sliced_traces = [traces[idx] for idx in take_idx]
-    return sliced_traces, final_length + tune
 
 
 def _init_jitter(
